@@ -1,7 +1,7 @@
+import { createRoot } from 'react-dom/client'
+import App from './App'
 import './index.css'
 
-import React from 'react'
-import ReactDOM from 'react-dom'
-import App from './App'
-
-ReactDOM.render(<App />, document.getElementById('root'))
+const container = document.getElementById('root')
+const root = createRoot(container!)
+root.render(<App />)
